task 6 , 7

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 """ Moudule """
-from typing import Union, List, Tuple
+from typing import Union, List, T
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """ Function """
     return (k, v ** 2)
